+++ test/integration_test.cpp
@@ -525,6 +525,79 @@ ctrl.X q[0], q[1]
     }));
 }
 
+TEST_F(IntegrationTest, control_gate_modifier__ctrl_x90) {
+    auto program = R"(
+version 3.0
+
+qubit[2] q
+
+H q[0]
+ctrl.X90 q[0], q[1]
+H q[0]
+H q[1]
+)";
+    std::size_t iterations = 1;
+    auto actual = run_from_string(program, iterations, "3.0");
+
+    // Expected 'q' state should be |00>+|10>+|11>
+    EXPECT_EQ(actual.state,
+        (SimulationResult::State{
+            { "00",                  core::Complex{ .real = 1. / gates::SQRT_2, .imag = 0, .norm = 0.5 } },
+            { "10",  core::Complex{ .real = gates::SQRT_2 / 4, .imag = gates::SQRT_2 / 4, .norm = 0.25 } },
+            { "11", core::Complex{ .real = gates::SQRT_2 / 4, .imag = -gates::SQRT_2 / 4, .norm = 0.25 } }
+    }));
+}
+
+TEST_F(IntegrationTest, control_gate_modifier__ctrl_x90_after_h_q1) {
+    auto program = R"(
+version 3.0
+
+qubit[2] q
+
+H q[1]
+ctrl.X90 q[0], q[1]
+H q[0]
+H q[1]
+)";
+    std::size_t iterations = 1;
+    auto actual = run_from_string(program, iterations, "3.0");
+
+    // Expected 'q' state should be |00>+|01>
+    EXPECT_EQ(actual.state,
+        (SimulationResult::State{
+            { "00", core::Complex{ .real = 1. / gates::SQRT_2, .imag = 0, .norm = 0.5 } },
+            { "01", core::Complex{ .real = 1. / gates::SQRT_2, .imag = 0, .norm = 0.5 } }
+    }));
+}
+
+TEST_F(IntegrationTest, control_gate_modifier__ctrl_rx_half_pi) {
+    auto program = R"(
+version 3.0
+
+qubit[2] q
+
+H q[0]
+ctrl.Rx(pi/2) q[0], q[1]
+H q[0]
+H q[1]
+)";
+    std::size_t iterations = 1;
+    auto actual = run_from_string(program, iterations, "3.0");
+
+    // Expected 'q' state should be |00>+|01>+|10>+|11>
+    EXPECT_EQ(actual.state,
+        (SimulationResult::State{
+            { "00",
+             core::Complex{ .real = (1. + gates::SQRT_2) / 4, .imag = -0.25, .norm = (2. + gates::SQRT_2) / 8 }      },
+            { "01",
+             core::Complex{ .real = (-1. + gates::SQRT_2) / 4, .imag = 0.25, .norm = (2. - gates::SQRT_2) / 8 }      },
+            { "10", core::Complex{ .real = (1. + gates::SQRT_2) / 4, .imag = 0.25, .norm = (2. + gates::SQRT_2) / 8 } },
+            { "11",
+             core::Complex{
+             .real = (-1. + gates::SQRT_2) / 4, .imag = -0.25, .norm = (2. - gates::SQRT_2) / 8 }                    }
+    }));
+}
+
 TEST_F(IntegrationTest, gate_modifier__ctrl_pow_2_s) {
     auto program = R"(
 version 3.0