+++ CHANGELOG.md
@@ -10,6 +10,16 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - **Removed** for now removed features.
 
 
+## [ 0.9.0 ] - [ 2025-04-07 ]
+
+### Added
+- Integrate with libqasm 1.2.0 release.
+
+### Fixed
+- Pi-half rotation gates defined incorrectly.
+- Bug in SparseComplex::operator=, which was not always performing the assignment.
+
+
 ## [ 0.8.0 ] - [ 2025-03-21 ]
 
 ### Added